Refactor PopulateStruct (#337)
config.PopulateStruct is not really recursive now: 
primitive types are available in full only for structs, 
but not arrays; maps can't have arrays or structs 
as values, etc. 

There is also some code duplication: the dispatch 
for populating values for an array, map, struct fields 
should be the same.

This commit splits valueStruct method in several 
smaller methods, that are called recursively. 
Which solves all the problems above.

Added support of arrays of fixed length
Check for cycles
Populate only exported fields