Add nop config provider (#409)
If users want to test a specific functionality, 
e.g. RegisterCallback always returns an error, 
the easiest way to do that will be embedding 
a NopConfigProvier, and they can't do this 
with a staticProvider, because it is private.

Another part is type preservation: in static provider
we serialize objects to yaml and loose type information, 
uints, int8s will be converted to ints, float32s to float64.