Check implementation of TextUnmarshaller on all types (#27)
Populate is using encoding.TextUnmarshaller interface only on scalar types,
functions and channels, but any type can implement it, e.g. zap.AtomicLevel

We are using Value.String() to construct the byte array. In case of scalar types
received from a provider, text does not need extra formatting. TextUnmarshaller
implementation should be aware of the source of the values for complex types.
For example in TestUnmarshalTextOnComplexStruct the actual contents of byte
array are map[year:1994 title:Free Willy]