No panics in config constructors (#43)
Instead of panicking on errors, we should simply return them
to a caller end let them decide what to do with the errors.

The change is mostly straight forward, but there is one snowflake:
expand variables. The replace function doesn't return an error,
so we panic internally and use recover to report errors.