resolve conflicts (#44)
(1) Remove the Command Line Provider

The command line provider is not be used in the vast majority of cases.

Removes the StringSlice type.

We can bring this functionality back if users request it, but the library will be easier to maintain without it.

(2) Un-export the Cache Provider

The cached provider is great for improving the performance of our YAML providers, but we don’t expect users to interact with it directly.

Users should be accessing a smaller set of exposed YAML provider solutions.

(3) Remove the config.Value.ChildKeys Method

The ChildKeys method is not used in any of the internal functions. In our customers’ experience, we have never needed to access a value’s child keys.
(4) Remove the config.Value.timestamp attribute and config.Value.LastUpdated Method

If we make this configuration library more focused on serving static YAML files, there is no need for maintaining timestamps in our values.

All internal references to config.NewValue passed nil for the value of the given timestamp, so removing the attribute was seemingly harmless.

We will not be reading updates so we can also remove the LastUpdated method.

(5) Update Godocs for Consistency

There are minor discrepancies between godocs on a variety of different methods. E.g. Comments not ending with periods.