Introduce a text.Transformer to perform variable expansion
Currently expansion of variables using a shell-like syntax like ${foo}
or $foo is supported, but it is performed on the unmarshalled data
structures _after_ YAML parsing.  A side-effect of this is that any
data type resolution that the YAML parser may perform is lost and the
resulting fields are all represented as strings.

Since the non-expanding YAML provider does not perform this filtering
step on the result of the YAML unmarshaller, it does not suffer from
this problem, but it does mean that the data structures constructed by
the non-expanding YAML provider differ from the expanding YAML provider.

Let's introduce a text.Transformer to perform the parsing of the
shell-like syntax and apply the expansion function on the variables.

This allows us to simplify the code paths in yaml.go and also to
define an escape sequence '$$' that can be used to avoid the variable
expansion when a literal '$foo' is desired.

Since we're no longer using os.Expand, we can modify the interface
of the function produced by replace() so that it can return an error on
failure instead of using panic/recover.

Update tests and add a few new ones.  This fixes the tests in
TestYamlProvidersProduceSameResults in yaml_test.go