Demonstrate that YAML providers produce different data types
Currently, the non-expanding YAML provider and the one that can expand
the shell-like variable syntax produce different output from the same
input even when no expansion is involved.  This is because the expanding
YAML provider converts all fields to string as a side-effect of its
operation.

Let's add some tests to demonstrate this incongruity.