yaml.go: don't leak file handles on error
Currently, yaml.go:unmarshalYAMLValue() is passed an io.ReadCloser and
assumes responsibility for closing the io.ReadCloser object that is
passed in.  If this function encounters an error, it will not close the
io.ReadCloser, and neither will any of its callers.  This could result
in a leak of file descriptors.

Let's modify the interface to unmarshalYAMLValue() and most callers so
that instead of being passed an io.ReadCloser(), or a slice of
io.ReadCloser()'s, and assuming the responsibility for closing them all,
we'll just pass them io.Reader()'s, and leave the responsibility for
closing any readers that need closing in the high-level callers that
opened the files in the first place, e.g. NewYAMLProviderFromFiles and
NewYAMLProviderWithExpand.

This has the additional benefit of simplifying some callers in
static_provider.go and tests which no longer need to wrap their objects
in ioutil.NopCloser().