Add an internal package for unreachable errors
Occasionally, we'll to handle an error case that shouldn't be
reachable. We still want to avoid panics and return errors where
possible, but we should also wrap the error message to clearly indicate
that this is an error on our part.

Note that since this commit introduces a new subpackage, we also need to
update our build tooling.