Add integration tests for NewValue (#81)
This is...unfortunate. Unless users are very careful, the behavior of
`NewValue` is completely irrational. These tests aren't exhaustive, but
the gist is this:

* `NewValue` takes a provider, a key, a value (as `interface{}`), and a
bool indicating whether the value was found or not. There's currently no
validation checking that `value` and `found` match the contents of the
provider at the specified key.
* `HasValue` returns the user-supplied bool, ignoring the contents of
the provider.
* `String` and `Value` return the user-supplied value, ignoring the
contents of the provider.
* `Populate`, `Get`, and `WithDefault` use the provider, ignoring the
user-supplied value and bool.

This leads to a lot of potentially counter-intuitive behavior. Since the
primary use case for `NewValue` is in tests, this is misleading - the
test may not reflect reality at all.

I'm opening this PR to let everyone ponder a good path forwards. There
are 90ish uses of `NewValue` internally, but very few of them use the
`Value` or `HasValue` APIs. I propose this: in `NewValue`, validate that
`value` reflects the contents of the provider at the specified key. If
it doesn't, panic (I know, I know). Ignore `found` completely. Once
validation is complete, just return `provider.Get(key)`. Before cutting
a new release, deprecate `NewValue` and many of the methods on `Value`,
including `Value()` and `HasValue()`.

This keeps most of the API compatible, makes the behavior sane-ish, and
isn't likely to break anyone's production code.