Replace provider and value implementations (#87)
With the foundation laid by an integration test suite and the internal
YAML-merging package, gut the internals of this package and reimplement
the core types.

The new implementation is intentionally naive: since we're only using
this package at process startup, we don't care much about performance
and can afford repeated round-trips through YAML. This has two benefits:
the code is simpler, and it's easier to leverage all the features that
`gopkg.in/yaml.v2` already has (especially strict unmarshalling mode).
While we have the opportunity, we also introduce a new constructor that
takes functional options. I've also improved the GoDoc and deprecated a
bunch of the existing APIs (though they all continue to function as
expected).

This fixes issues #57, #59, #72, #73, #75, #76, #80, and a pile of other
bugs that weren't reported by users.