YAML: support non-string scalar keys (#105)
YAML mappings can be keyed by any kind of node so long as it is unique.
That means a mapping key can be an integer, a boolean, or a number of
other types besides a string.

`Get` retrieves a value on a period-separated path and previously
interpreted each segment as a string. This change unmarshals each path
segment to produce comparable map keys, even if it's a bit strange when
"no: car" content produces a value with `provider.Get("false")`.